rebase, introduce Silent error and move ConstEvalErr to middle
