+++ compiler/rustc_monomorphize/src/collector.rs
@@ -731,6 +731,10 @@ impl<'a, 'tcx> MirVisitor<'tcx> for MirNeighborCollector<'a, 'tcx> {
                             "collection encountered polymorphic constant: {:?}",
                             literal
                         ),
+                        Err(ErrorHandled::Silent(_)) => span_bug!(
+                            self.body.source_info(location).span,
+                            "silent error emitted during collection",
+                        ),
                     }
                 }
                 _ => return,
@@ -762,6 +766,10 @@ impl<'a, 'tcx> MirVisitor<'tcx> for MirNeighborCollector<'a, 'tcx> {
                         self.body.source_info(location).span,
                         "collection encountered polymorphic constant: {}",
                         substituted_constant
+                    ),
+                    Err(ErrorHandled::Silent(_)) => span_bug!(
+                        self.body.source_info(location).span,
+                        "silent error emitted during collection",
                     ),
                 }
             }
